+++ pkg/config/config_test.go
@@ -2,6 +2,7 @@ package config
 
 import (
 	"flag"
+	"net/url"
 	"os"
 	"strings"
 	"testing"
@@ -10,6 +11,7 @@ import (
 	"github.com/grafana/agent/pkg/metrics"
 	"github.com/grafana/agent/pkg/metrics/instance"
 	"github.com/grafana/agent/pkg/util"
+	commonCfg "github.com/prometheus/common/config"
 	"github.com/prometheus/common/model"
 	promCfg "github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/pkg/labels"
@@ -427,4 +429,26 @@ metrics:
 	require.Equal(t, "verysecret", cfg.Metrics.ServiceConfig.KVStore.Etcd.Password)
 	require.Equal(t, "verysecret", cfg.Metrics.ServiceConfig.Lifecycler.RingConfig.KVStore.Consul.ACLToken)
 	require.Equal(t, "verysecret", cfg.Metrics.ServiceConfig.Lifecycler.RingConfig.KVStore.Etcd.Password)
+}
+
+func TestConfig_RemoteWriteDefaults(t *testing.T) {
+	cfg := `
+metrics:
+  global:
+    remote_write:
+      - name: "foo"
+        url: "https://test/url"`
+
+	var c Config
+	err := LoadBytes([]byte(cfg), false, &c)
+	require.NoError(t, err)
+
+	expected := &promCfg.DefaultRemoteWriteConfig
+	expected.Name = "foo"
+	testURL, _ := url.Parse("https://test/url")
+	expected.URL = &commonCfg.URL{
+		URL: testURL,
+	}
+	require.Equal(t, expected, c.Metrics.Global.RemoteWrite[0])
+	require.True(t, c.Metrics.Global.RemoteWrite[0].SendExemplars)
 }