+++ pkg/integrations/v2/autoscrape/autoscrape_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/prometheus/prometheus/discovery"
 	"github.com/prometheus/prometheus/pkg/exemplar"
 	"github.com/prometheus/prometheus/pkg/labels"
-	"github.com/prometheus/prometheus/scrape"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -98,11 +97,8 @@ func (ma *mockAppender) AppendExemplar(ref uint64, l labels.Labels, e exemplar.E
 }
 
 type mockInstance struct {
+	instance.NoOpInstance
 	app storage.Appender
 }
 
-func (mi *mockInstance) Run(ctx context.Context) error                 { return nil }
-func (mi *mockInstance) Update(c instance.Config) error                { return nil }
-func (mi *mockInstance) TargetsActive() map[string][]*scrape.Target    { return nil }
-func (mi *mockInstance) StorageDirectory() string                      { return "" }
 func (mi *mockInstance) Appender(ctx context.Context) storage.Appender { return mi.app }